Добавление загруженного фото в форму
