+++ js/form.js
@@ -2,27 +2,57 @@ import { resetEffects } from './effects.js';
 import { resetScale } from './scale.js';
 import { sendData } from './api.js';
 import { showErrorMessage, showSuccesMessage } from './message.js';
+import { initValidation, pristineValid, resetValidator } from './validation.js';
 
-const MAX_COMMENT_LENGTH = 140;
-const MAX_HASHTAG_COUNT = 5;
-const VALID_HASHTAG = /^#[a-zа-яё0-9]{1,19}$/i;
+
 const SUBMIT_BUTTON_TEXT = {
   default: 'Опубликовать',
   sending: 'Отправка...'
 };
 
+const FILE_TYPES = ['jpg', 'jpeg', 'png'];
+
+const uploadButtonElement = document.querySelector('#upload-file');
 const imgUploadForm = document.querySelector('.img-upload__form');
 const imgUploadOverlay = document.querySelector('.img-upload__overlay');
 const imgUploadCancel = document.querySelector('.img-upload__cancel');
 const imgUploadSubmit = document.querySelector('.img-upload__submit');
 const textHashtags = document.querySelector('.text__hashtags');
 const textDescription = document.querySelector('.text__description');
+const imagePreviewElement = document.querySelector('.img-upload__preview img');
+const effectsPreviewElement = document.querySelectorAll('.effects__preview');
 const body = document.body;
 
-const pristine = new Pristine(imgUploadForm, {
-  classTo: 'img-upload__field-wrapper',
-  errorTextParent: 'img-upload__field-wrapper',
+imgUploadForm.addEventListener('submit', (evt) => {
+  evt.preventDefault();
+  pristineValid();
 });
+
+const showForm = () => {
+  resetValidator();
+  imgUploadOverlay.classList.remove('hidden');
+  body.classList.add('modal-open');
+  document.addEventListener('keydown', onDocumentKeydown);
+
+};
+
+const closeForm = () => {
+  resetEffects();
+  resetScale();
+  imgUploadForm.reset();
+  imgUploadOverlay.classList.add('hidden');
+  body.classList.remove('modal-open');
+  document.removeEventListener('keydown', onDocumentKeydown);
+  resetValidator();
+};
+
+function onDocumentKeydown(evt) {
+  const isInputFocused = [textHashtags, textDescription].some((el) => el === evt.target);
+  if (evt.key === 'Escape' && !isInputFocused){
+    evt.preventDefault();
+    closeForm();
+  }
+}
 
 const blockSubmitButton = () => {
   imgUploadSubmit.disabled = true;
@@ -35,10 +65,11 @@ const unblockSubmitButton = () => {
 };
 
 const setUserFormSubmit = (onSuccess) => {
+  initValidation();
   imgUploadForm.addEventListener('submit', (evt) => {
     evt.preventDefault();
 
-    if (pristine.validate()) {
+    if (pristineValid()) {
       blockSubmitButton();
       sendData(new FormData(evt.target))
         .then(() => {
@@ -51,87 +82,26 @@ const setUserFormSubmit = (onSuccess) => {
         .finally(unblockSubmitButton);
     }
   });
+  uploadButtonElement.addEventListener('change', () => {
+    const file = uploadButtonElement.files[0];
+    const fileName = file.name.toLowerCase();
+    const matches = FILE_TYPES.some((extension) => fileName.endsWith(extension));
+    if (matches) {
+      imagePreviewElement.src = URL.createObjectURL(file);
+      effectsPreviewElement.forEach((effect) => {
+        effect.style.backgroundImage = `url('${URL.createObjectURL(file)}')`;
+      });
+      imgUploadForm.addEventListener('change', showForm);
+    } else {
+      imgUploadForm.removeEventListener('change', showForm);
+      showErrorMessage();
+      closeForm();
+    }
+  });
 };
-
-function validateComment (value) {
-  return value.length <= MAX_COMMENT_LENGTH;
-}
-
-pristine.addValidator(
-  imgUploadForm.querySelector('.text__description'),
-  validateComment,
-  'Длина комментария не может составлять больше 140 символов'
-);
-
-const splitHashtags = (hashtags) => hashtags.trim().split(/\s+/);
-
-function validateHashtagItems (value) {
-  const hashtags = splitHashtags(value);
-  const isValidCount = hashtags.length <= MAX_HASHTAG_COUNT;
-  const isValidText = hashtags.every((hashtag) => VALID_HASHTAG.test(hashtag));
-  const isUnique = hashtags.length === new Set(hashtags.map((hashtag) => hashtag.toLowerCase())).size;
-
-  return {isValidCount, isValidText, isUnique};
-}
-
-function validateHashtag (value) {
-  const {isValidCount, isValidText, isUnique} = validateHashtagItems(value);
-  return isValidCount && isValidText && isUnique;
-}
-
-const getHashtagErrorMessage = (value) => {
-  const {isValidCount, isValidText, isUnique} = validateHashtagItems(value);
-
-  if (!isValidCount) {
-    return 'Нельзя указать больше пяти хэш-тегов';
-  }
-  if (!isValidText) {
-    return 'Строка после решётки должна состоять из букв и чисел и иметь длину не более 20 символов';
-  }
-  if (!isUnique) {
-    return 'Один и тот же хэш-тег не может быть использован дважды';
-  }
-  return true;
-};
-
-pristine.addValidator(
-  imgUploadForm.querySelector('.text__hashtags'),
-  validateHashtag,
-  getHashtagErrorMessage
-);
-
-imgUploadForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  pristine.validate();
-});
-
-
-const showForm = () => {
-  imgUploadOverlay.classList.remove('hidden');
-  body.classList.add('modal-open');
-  document.addEventListener('keydown', onDocumentKeydown);
-};
-
-const closeForm = () => {
-  resetEffects();
-  resetScale();
-  imgUploadForm.reset();
-  imgUploadOverlay.classList.add('hidden');
-  body.classList.remove('modal-open');
-  document.removeEventListener('keydown', onDocumentKeydown);
-};
-
-function onDocumentKeydown(evt) {
-  const isInputFocused = [textHashtags, textDescription].some((el) => el === evt.target);
-  if (evt.key === 'Escape' && !isInputFocused){
-    evt.preventDefault();
-    closeForm();
-  }
-}
-
 
 imgUploadCancel.addEventListener('keydown', onDocumentKeydown);
 imgUploadCancel.addEventListener('click', closeForm);
-imgUploadForm.addEventListener('change', showForm);
+
 
 export { setUserFormSubmit, closeForm };
